p_set: discard any repetitive pattern in the rotation set
For example, if we create the rotation set (8 positions) of the square, all 8
matrix will be identical. Add a check to add in the rotation set only the new
patterns.

With the previous example, we have a performance increase of ~210% (3x faster)

➜  time python solver.py
---
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 42 42 0  0  0  0
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 42 42 0  0  0  0
42 42 42 42 42 42 0  0  0  0  0
---
42 42 42 42 42 42 12 12 2  2  2
42 42 42 42 42 42 42 12 2  8  10
42 42 42 42 42 42 12 12 8  8  10
42 42 42 42 42 42 42 8  8  10 10
42 42 42 42 42 42 9  9  9  9  10
python solver.py  3,54s user 0,29s system 113% cpu 3,384 total