p_set: compute the rotation set only once
For each piece, compute the whole rotation set (without duplicates) at start,
and store it. Instead of computing it each time we place a piece (which lead to
numerous identical computation, as the same piece may be placed, then removed,
the re-placed, and so on).

This gives ~164% performance increase (x2.5 faster)

➜  time python solver.py
---
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 42 42 0  0  0  0
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 42 42 0  0  0  0
42 42 42 42 42 42 0  0  0  0  0
---
42 42 42 42 42 42 12 12 2  2  2
42 42 42 42 42 42 42 12 2  8  10
42 42 42 42 42 42 12 12 8  8  10
42 42 42 42 42 42 42 8  8  10 10
42 42 42 42 42 42 9  9  9  9  10
python solver.py  1,34s user 0,10s system 107% cpu 1,339 total