p_set: dynamic piece shape
Previous method fixed the matrix size for each piece & rotations to 4x4,
aligning the piece to the 'top-left' corner of the matrix when computing a new
rotation. The grid was also extended by 3 rows and 3 columns, to allow the
comparison of a piece matrix with the far-right and/or far-bottom of the grid,
without trespassing the matrix boundaries.
For the vast majority of the pieces, this caused an check of a few impossible
positions (with part of the piece 'outside' the grid matrix) because we had to
support the few cases of very thin/small pieces.

Changing this to use dynamic size: each piece matrix has its own size, that we
can access via the '.shape' property. The computation of each rotation is also
easy, as the 'np.rot90()' function supports it well.

Removing the additional lines of the grid matrix.

We have a performance increase of ~77% (x1.7 faster) with the latest control grid.

➜  time python solver.py
---
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
---
42 42 42 42 42 42 11 11 7  7  8
42 42 42 42 42 10 11 11 7  8  8
42 42 42 4  10 10 10 10 8  8  3
42 42 42 4  9  9  9  9  2  3  3
42 42 42 4  4  4  2  2  2  3  3
python solver.py  16,51s user 0,24s system 100% cpu 16,673 total