solver: compute nb of connected components to reduce false tracks
After placing a new piece, compute the number of places that are left (using a
DFS). If there are more places to be filled than the number of pieces, it means
that the current place is wrong.

With this, we have a performance increase of ~347% (x4.5 faster)

➜  time python solver.py
---
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
---
42 42 42 42 42 42 11 11 7  7  8
42 42 42 42 42 10 11 11 7  8  8
42 42 42 4  10 10 10 10 8  8  3
42 42 42 4  9  9  9  9  2  3  3
42 42 42 4  4  4  2  2  2  3  3
python solver.py  3,39s user 0,33s system 111% cpu 3,330 total