p_set: cancel attempt if it leave places too small for any piece
In the current set of pieces, no piece can fit in a place of size of 2 or less.
Cancel this attempts early without finishing to count the number of places left.

This gives a performance increase of ~334% (x4.3 faster) with the control grid

➜  time python solver.py
---
42 42 42 42 42 42 0  0  0  0  0
42 42 42 42 42 0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
42 42 42 0  0  0  0  0  0  0  0
---
42 42 42 42 42 42 11 11 7  7  8
42 42 42 42 42 10 11 11 7  8  8
42 42 42 4  10 10 10 10 8  8  3
42 42 42 4  9  9  9  9  2  3  3
42 42 42 4  4  4  2  2  2  3  3
python solver.py  0,78s user 0,28s system 165% cpu 0,639 total